responding to a few comments
