+++ src/graphql/data/nft/Details.ts
@@ -1,12 +1,10 @@
 import { parseEther } from '@ethersproject/units'
 import graphql from 'babel-plugin-relay/macro'
 import { CollectionInfoForAsset, GenieAsset, SellOrder, TokenType } from 'nft/types'
-import { useEffect, useMemo, useState } from 'react'
-import { fetchQuery, useLazyLoadQuery, useQueryLoader, useRelayEnvironment } from 'react-relay'
-import { Subscription } from 'relay-runtime'
+import { useEffect } from 'react'
+import { useLazyLoadQuery, useQueryLoader } from 'react-relay'
 
 import { DetailsQuery } from './__generated__/DetailsQuery.graphql'
-import { DetailsSellOrdersQuery } from './__generated__/DetailsSellOrdersQuery.graphql'
 
 const detailsQuery = graphql`
   query DetailsQuery($address: String!, $tokenId: String!) {
@@ -87,43 +85,6 @@ const detailsQuery = graphql`
           traits {
             name
             value
-          }
-        }
-      }
-    }
-  }
-`
-
-const sellOrdersQuery = graphql`
-  query DetailsSellOrdersQuery($address: String!, $tokenId: String!) {
-    nftAssets(address: $address, filter: { tokenIds: [$tokenId] }) {
-      edges {
-        node {
-          listings(first: 10) {
-            edges {
-              node {
-                address
-                createdAt
-                endAt
-                id
-                maker
-                marketplace
-                marketplaceUrl
-                orderHash
-                price {
-                  currency
-                  value
-                }
-                quantity
-                startAt
-                status
-                taker
-                tokenId
-                type
-                protocolParameters
-              }
-              cursor
-            }
           }
         }
       }
@@ -219,51 +180,4 @@ export function useDetailsQuery(address: string, tokenId: string): [GenieAsset, 
       externalUrl: collection?.homepageUrl ?? undefined,
     },
   ]
-}
-
-export function useSellOrdersQuery(
-  address: string,
-  tokenId: string,
-  enabled: boolean
-): { sellOrders: SellOrder[] | undefined; isLoading: boolean } {
-  const environment = useRelayEnvironment()
-  const [sellOrders, setSellOrders] = useState<SellOrder[] | undefined>()
-  const [isLoading, setIsLoading] = useState(true)
-
-  useEffect(() => {
-    let subscription: Subscription | undefined
-
-    if (enabled) {
-      subscription = fetchQuery<DetailsSellOrdersQuery>(environment, sellOrdersQuery, {
-        address,
-        tokenId,
-      }).subscribe({
-        next(data) {
-          const asset = data.nftAssets?.edges[0]?.node
-          const map = asset?.listings?.edges.map((listingNode) => {
-            return {
-              ...listingNode.node,
-              protocolParameters: listingNode.node.protocolParameters
-                ? JSON.parse(listingNode.node.protocolParameters.toString())
-                : undefined,
-            } as SellOrder
-          })
-          setSellOrders(map)
-        },
-        complete: () => setIsLoading(false),
-      })
-    }
-
-    return () => {
-      if (subscription) subscription.unsubscribe()
-    }
-  }, [address, enabled, environment, tokenId])
-
-  return useMemo(
-    () => ({
-      sellOrders,
-      isLoading,
-    }),
-    [sellOrders, isLoading]
-  )
 }
