+++ src/nft/components/collection/CollectionAsset.tsx
@@ -5,7 +5,7 @@ import { sendAnalyticsEvent, useTrace } from '@uniswap/analytics'
 import { EventName, PageName } from '@uniswap/analytics-events'
 import { MouseoverTooltip } from 'components/Tooltip'
 import Tooltip from 'components/Tooltip'
-import { useSellOrdersQuery } from 'graphql/data/nft/Details'
+import { useSellOrdersQuery } from 'graphql/data/nft/SellOrders'
 import { Box } from 'nft/components/Box'
 import { bodySmall } from 'nft/css/common.css'
 import { useBag } from 'nft/hooks'
